RowOrScalar refactor
