+++ crates/dataflow-jit/src/ir/graph.rs
@@ -460,7 +460,8 @@ mod tests {
             graph::{Graph, GraphExt},
             literal::{NullableConstant, RowLiteral},
             nodes::{Differentiate, Fold, IndexWith, Neg, Source, StreamLayout, Sum},
-            types::{ColumnType, RowLayout, RowLayoutBuilder},
+            row_layout::{RowLayout, RowLayoutBuilder},
+            types::ColumnType,
             validate::Validator,
         },
         row::{Row, UninitRow},
